gimp_message. libgimp also overrides g_message for all plugins. Converted
* redid the error message handling. g_message now calls message_box or prints
to console depending on whether the no_interface is set or not. gimp-message
is also exported to the PDB as a wrapper to g_message, and libgimp has a new
API: gimp_message. libgimp also overrides g_message for all plugins. Converted
lots of messages in app/* to g_message. Made script-fu a little friendlier.

* updated the regex code from grep 2.2

* said goodbye to the old script-fu logo in script-fu.h

-Yosh


Former-commit-id: 1d95a05af0ddb236d2beece2e6a1ab4e3d127d45