Put in a workaround for the problem with plug-ins on NT and W2K, where
1999-08-28  Tor Lillqvist  <tml@iki.fi>

* libgimp/gimp.c (gimp_main): Put in a workaround for the problem
with plug-ins on NT and W2K, where installing the GIMP under a
directory path with spaces (\Program Files\GIMP) would cause the
plug-ins to malfunction. Turns out the argv gets set up by the C
runtime incorrectly in such a case, the path name of the
executable is split at the spaces. The workaround is to splice
argv[0] together from the pieces again, shifting the rest of argv
down.

* app/plug_in.c (plug_in_open): Correct (harmless) typo in the
Win32 code.

* plug-ins/common/winclipboard.c: Put the pasting of a new image
under File/Acquire.


Former-commit-id: 3e2ef6680900676498dace2e6aaf319e7d7dda19