libgimp/gimpenv.c (gimp_path_parse) Don't try to handle tilde (~) at all
2000-05-16  Tor Lillqvist  <tml@iki.fi>

* libgimp/gimpenv.c (gimp_path_parse)
* app/gimprc.c (transform_path): Don't try to handle tilde (~) at
all on Windows. The tilde is used when mangling long file names
into 8.3 form, and some people have a HOME environment variable
that contains this kind of 8.3 path with a tilde. This causes
interesting effects if we expand a tilde in some gimprc variable
with $HOME, which contains a tilde, which we expand with $HOME,
etc.


Former-commit-id: bc1db67d83cfdd7f718349015825d6b89a836f3a