Made the explicit target "regenerate" to regenerate the source files using
* app/composite/Makefile.am:
Made the explicit target "regenerate" to regenerate the source
files using the Python meta-generator.

* app/composite/gimp-composite-mmx.c
* app/composite/gimp-composite-mmx.h
* app/composite/gimp-composite-sse.c
* app/composite/gimp-composite-sse.h:
Now bounded by the conditional "#if __GNUC__ >= 3".  This is a
coarse fix to enable app/composite to build on a wider variety of
platforms.  Ideally, this will be refined to encompass different
versions of the compiler(s) supporting inline assembly code.


Former-commit-id: 0aad627dbe2f7a61c07f62b7e203a1b8a2544e58