Avoid use of temporary files for accessing remote files
When using GIO/GVFS, remote files may have a local representation
provided by a FUSE mount. Try to detect this and access the file
directly instead of using a temporary file. Fallback to the old
behavior of using temporary files if GIO can't give us a local path
to the file.


Former-commit-id: 6fa2ece547efa1ad515648419f1444ce0b1e948a