Revert "Only didive if the number of inputs is > 1 (/ 1 makes little sense)"
This reverts commit 4a56e6799171979da82dbcb9281301bdca3626ad [formerly 793be22da944b02e67ad4c6ad7b40afd70184889].

This commit makes single factor dynamics not work.
The result of a no-factor mixing needs to be 1.0,
result of all other cases needs to be value of total divided by factors.
Adding extra logic for the case whe factors=1 makes less sense than
allowing for occational division by 1 to happen.


Former-commit-id: a7234d68ce468bc92baaaa032f882a7dc1eb4b66