Modified UI to include default timing
This patch makes two modifications to the UI: listing in the
'Playback speed' combo was changed from percentage values to X
factors, and a new combo was added for 'Default framerate'. The
latter includes common framerates which are used to define the
default duration on unmarked frames. Due to the way in which the
GTK+ timeout function is defined, the framerates and duration values
for this new combo use rounded integer values.


Former-commit-id: 85b8d9668e8abccad2fe9015f6fd75d383ea6276