devel-docs: Add a schedule for development
Add an OpenOffice.org Calc document that contains a list of tasks that
needs to be done before GIMP 2.8 is ready for release. An estimate of
the time to complete each task, in 8-hour work days, is also
given. Based on this data an estimated date when we will have a GIMP
2.8 release candidate ready is calculated. The formula includes a
“days worked per week”-factor that specifies how many 8-hour work days
that the GIMP community together produces per week. There is also a
sheet where things planned for GIMP 2.10/3.0 is put. With this
document we will be able to better plan what features to include in
what version.


Former-commit-id: 038ea9474d9f0fe6877376d1cc3640b1bde8d3b8