plug-ins: Change MNG capability from LAYERS_AS_ANIMATION to LAYERS
When NULL is passed as format_name to gimp_export_image(), there is no
dialog where the user can make choices. Having the capability
GIMP_EXPORT_CAN_HANDLE_LAYERS_AS_ANIMATION in this case in practice
means "always merge". So to allow MNG animations, change to just
GIMP_EXPORT_CAN_HANDLE_LAYERS. We don't need to offer merging in the
export options since MNG is a format explicitly for animation, so if
there are enough layers, we'll always export as an animation.

Also set a tooltip explaining why the animation options are disabled
when they are disabled.


Former-commit-id: ac1baa5374b92423d618dbba46dd97addf9a7fd7